Create an "About" tab to explain this tool.
Amazon Research Description is something like "You have to enter the segment that you want to research on amazon, that will go as the search text on amazon. The tool will then scrape top 7 pages for that keyword and store the data in the google sheets (the link for which will be mentioned in the mail sent by us). The tool will then do its analysis and generate a report which will be sent to your mail (within 5 minutes) with the data link attached. You can only generate reports in a month.

Flipkart Scrapper Description is something like: "You have to enter list of FSN either through copying from your excel or by comma separated. The tool will then scrape those product urls (the link will be sent to you through mail). The important product details like title, price, rating, rating count etc will be stored in google sheets. A mail will be sent to you with the data source link within 5 minutes. Please keep the list of FSN upto 80.

You can reframe the text and beautify it in the "About" section tab. Make it easy to read and attractive, and comprehensive.